requirements.txt
flask
flask-cors
flaskwebgui
numpy
PyPDF2
pandas
python-pptx
pytesseract
Pillow
torch
concurrent-log-handler
faiss-cpu
transformers
sentence-transformers
openpyxl